+++ sourcecode.txt
@@ -1,6 +1,5 @@
 program ( )
 {
-	declare list of variables a b c : array [ 90 .. 30 ] [ 70 .. 90 ] [ 3 .. 8 ] of integer ;
 	declare list of variables abcd ghd2_1 ssd_2_3 : array [ 2 .. 5 ] of integer ;
 	declare list of variables new d w e2 : integer ;
 	declare v1 : integer ;
@@ -8,9 +7,18 @@ program ( )
 	declare list of variables b1 b2 b3 : boolean ;
 	declare list of variables c1 c2 : real ;
 	declare list of variables u v : array [ 2 .. 5 ] [ 3 .. 6 ] of  integer ;
-	declare list of variables s1 s2 s3 :  jagged array [ 4 .. 7 ] [ ] [ ] of integer ;
-	R1 [ 4 ]  : size 3 : values { 21 641 23 36 125 ; 54 221 43 }
+	declare list of variables s1 s2 s3 :  jagged array [ 4 .. 7 ] [ ] [ ]  of integer ; 
+	R1 [ 4 ]  : size 3 : values { 21 641 23 36 125 ; 54 221 43 ; 287 501 453 334 23 }
 	R1 [ 5 ]  : size 2 : values { 12 10 100 ; 76 15 8 54 432 }
 	R1 [ 6 ]  : size 3 : values { 17 61 928 785 875 ; 334 121 61 9 ; 32 465 123 } 
 	R1 [ 7 ]  : size 4 : values { 210 71 ; 90 47 32 10 93 ; 332 453  12 634 ; 44 53 55 134 }
+	declare list of variables m1 m2 m3 : real ;
+	declare list of variables h1 h2 h3 : array [ 12 .. 34 ] [ 11 .. 20 ] [ 18 .. 33 ] of integer ;
+	declare list of variables s4 s5 s6 : jagged array [ 3 ..  8 ] [ ] of integer ; 
+	R1 [ 3 ]  : size 3 : values { 20 ; 35 ; 54 }
+	R1 [ 4 ]  : size 6 : values { 65 ; 89 ; 99 ; 11 ; 37 ; 11 }
+	R1 [ 5 ]  : size 2 : values { 22 ; 745 }
+	R1 [ 6 ]  : size 4 : values { 67 ; 91 ; 13 ; 44 }
+	R1 [ 7 ]  : size 1 : values { 17 }
+	R1 [ 8 ]  : size 5 : values { 31 ; 97 ; 10 ; 9 ; 120 }
 }
